# Speed up fastscan package